Move Board tests into a separate module
